Fix tests
Text::CSV_XS defaults `decode_utf8` to true, so the tests were failing
because utf8 strings were compared with byte strings. Another fix was
to just add `use utf8` to the beginning of the file. But I think that
explicit disabling of `decode_utf8` is closer in spirit to what was
intended.